requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
fpdf>=2.7.0
scikit-learn>=1.3.0
scipy>=1.8.0
